refactor(channels/discord): P2a — extract markers.rs (egress trust boundary)
Move the outbound media-marker subsystem into discord/markers.rs: marker
parsing, the validate_marker_target trust boundary (http(s)+absolute-workspace
only), attachment classification, and the delivery-failure note/reactions.
marker_kind_for stays in mod.rs (inbound). Re-exported so callers + tests
resolve unchanged. Behaviour-neutral: 157 marker/classify/attachment tests pass.
